+++ Schema/Scripts/PostgreSQL/SchemaSmith.IndexOnlyQuench.sql
@@ -201,11 +201,11 @@ BEGIN
                                      ELSE 'UNIQUE ' || CASE WHEN ti."NullsNotDistinct" THEN 'NULLS NOT DISTINCT ' ELSE '' END
                                      END ||
                                 '(' || "SchemaSmith"."QuoteIndexColumnList"(ti."IndexColumns") || ')' ||
-                                CASE WHEN ti."Deferrable" THEN ' DEFERRABLE' ELSE '' END ||
-                                CASE WHEN ti."InitiallyDeferred" THEN ' INITIALLY DEFERRED' ELSE '' END ||
                                 CASE WHEN COALESCE(ti."AccessMethod", 'btree') NOT IN ('gin', 'brin', 'spgist')
                                      THEN ' WITH (fillfactor = ' || ti."FillFactor" || ')'
-                                     ELSE '' END || ';'
+                                     ELSE '' END ||
+                                CASE WHEN ti."Deferrable" THEN ' DEFERRABLE' ELSE '' END ||
+                                CASE WHEN ti."InitiallyDeferred" THEN ' INITIALLY DEFERRED' ELSE '' END || ';'
                            ELSE 'CREATE ' || CASE WHEN ti."Unique" THEN 'UNIQUE ' ELSE '' END || 'INDEX "' || ti."Name" || '" ON "' || ti."TableSchema" || '"."' || ti."TableName" || '" ' ||
                                 'USING ' || COALESCE(ti."AccessMethod", 'btree') || ' ' ||
                                 '(' || "SchemaSmith"."QuoteIndexColumnList"(ti."IndexColumns") || ')' ||